refactor: 알림 설정이 활성화됐을 때만 알림이 오도록 변경
